Moved make_hashable() to django.utils and added tests.
